Variable declarators working
